Split dataset tables tab content into own component
